requirements.txt
python-telegram-bot==20.7
requests==2.31.0
tweepy==4.14.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
pytest==7.4.3
pytest-asyncio==0.21.1
